requirements.txt
absl-py==0.15.0
ale-py==0.8.1
astunparse==1.6.3
atari-py==0.2.9
AutoROM==0.6.1
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
charset-normalizer==3.4.1
clang==5.0
click==8.1.8
cloudpickle==1.3.0
decorator==5.2.1
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.23
etils==1.3.0
flatbuffers==1.12
future @ file:///home/conda/feedstock_root/build_artifacts/future_1708610096684/work
gast==0.3.3
glfw==2.9.0
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
gym @ file:///weka/scratch/enalisn1/ziyao/dreamer-calibration-experiments/gym-0.19.0
gym-notices==0.0.8
h5py==2.10.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
keras==2.13.1
Keras-Preprocessing==1.1.2
labmaze==1.0.6
libclang==18.1.1
lxml==5.3.2
Markdown==3.7
MarkupSafe==2.1.5
mujoco==3.2.3
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649281352817/work
oauthlib==3.2.2
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==24.2
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyglet==1.5.0
PyOpenGL==3.1.9
pyparsing==3.1.4
requests-oauthlib==2.0.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
scipy==1.10.1
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.2
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.12.2
termcolor==1.1.0
tqdm==4.67.1
typing-extensions==3.7.4.3
unrar @ file:///home/conda/feedstock_root/build_artifacts/unrar_1674196625784/work
Werkzeug==3.0.6
wrapt==1.12.1
zipp==3.20.2
